ansible/roles/cron/node-discovery/files/node-discovery/requirements.txt
slackclient == 2.9.3
